% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairUpdate.R
\name{appendMFAIR}
\alias{appendMFAIR}
\title{Append the fitted factor to the MFAIR object in the greedy algorithm.}
\usage{
appendMFAIR(object, object_sf)
}
\arguments{
\item{object}{An MFAIR object containing the information about the first few factors.}

\item{object_sf}{An MFAIRSingleFactor needed to be appended.}
}
\value{
MFAIR object containing the information about the new fitted single factor.
}
\description{
Append the fitted factor to the MFAIR object in the greedy algorithm.
}
