% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfai_object.R
\name{createMFAI}
\alias{createMFAI}
\title{Create the MFAI object with main data matrix and auxiliary information.}
\usage{
createMFAI(Y, X, project = "MFAI")
}
\arguments{
\item{Y}{numeric. The main data matrix.}

\item{X}{data.frame. The auxiliary information.}

\item{project}{Name of the project (for record keeping).}
}
\value{
Returns MFAI object, with main data matrix and auxiliary information.
}
\description{
Create the MFAI object with main data matrix and auxiliary information.
}
