% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairObject.R
\name{createMFAIR}
\alias{createMFAIR}
\title{Create the MFAIR object with main data matrix and auxiliary information.}
\usage{
createMFAIR(Y, X, project = "MFAIR")
}
\arguments{
\item{Y}{numeric. The main data matrix.}

\item{X}{data.frame. The auxiliary information.}

\item{project}{Name of the project (for record keeping).}
}
\value{
Returns MFAIR object, with main data matrix and auxiliary information.
}
\description{
Create the MFAIR object with main data matrix and auxiliary information.
}
