% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairBackfitting.R
\name{fitBack}
\alias{fitBack}
\title{Fit the MFAI model using greedy algorithm.}
\usage{
fitBack(
  object,
  learning_rate = 0.1,
  minsplit = 10,
  minbucket = round(minsplit/3),
  maxdepth = 2,
  iter_max_bf = 5000,
  tol_bf = 0.01,
  verbose_bf_inner = TRUE,
  verbose_bf_outer = TRUE,
  ...
)
}
\arguments{
\item{object}{MFAIR object.}

\item{learning_rate}{Numeric. Parameter for the gradient boosting part.}

\item{minsplit}{Numeric. Parameter for the gradient boosting part.}

\item{minbucket}{Numeric. Parameter for the gradient boosting part.}

\item{maxdepth}{Numeric. Parameter for the gradient boosting part.}

\item{iter_max_bf}{Integer. Maximum iterations allowed.}

\item{tol_bf}{Numeric. The convergence criterion.}

\item{verbose_bf_inner}{Logical. Whether to display the detailed information during the inner loop.}

\item{verbose_bf_outer}{Logical. Whether to display the detailed information during the outer loop.}

\item{...}{}
}
\value{
An MFAIR object containing the information about the fitted MFAI model using backfitting algorithm.
}
\description{
Fit the MFAI model using greedy algorithm.
}
