% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairFitting.R
\name{fitGreedy,MFAIR-method}
\alias{fitGreedy,MFAIR-method}
\title{Greedy algorithm for MFAIR object.}
\usage{
\S4method{fitGreedy}{MFAIR}(object, ...)
}
\arguments{
\item{MFAIR}{MFAIR object.}
}
\value{
MFAIR object containing the information about the fitted MFAI model using greedy algorithm.
}
\description{
Greedy algorithm for MFAIR object.
}
