% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairGreedy.R
\name{fitGreedy}
\alias{fitGreedy}
\title{Fit the MFAI model using greedy algorithm.}
\usage{
fitGreedy(
  object,
  K_max = NULL,
  learning_rate = 0.1,
  minsplit = 10,
  minbucket = round(minsplit/3),
  maxdepth = 2,
  other_tree_para = list(),
  tol_snr = 0.002,
  verbose_greedy = TRUE,
  save_init = FALSE,
  sf_para = list()
)
}
\arguments{
\item{object}{MFAIR object.}

\item{K_max}{Integer. The maximum rank allowed in the MFAI model.}

\item{learning_rate}{Numeric. Parameter for the gradient boosting part.}

\item{minsplit}{Integer. Parameter for the gradient boosting part.}

\item{minbucket}{Integer. Parameter for the gradient boosting part.}

\item{maxdepth}{Integer. Parameter for the gradient boosting part.}

\item{other_tree_para}{A list containing other parameters for the gradient boosting part. See rpart::rpart.control() for details.}

\item{tol_snr}{Numeric. The convergence criterion which determine the inferred rank of data.}

\item{verbose_greedy}{Logical. Whether to display the detailed information when fitting the model.}

\item{save_init}{Logical. Whether to save the initialization of the model.}

\item{sf_para}{A list containing parameters for fitting the single factor MFAI model. See fitSFFully(), fitSFMissing(), or fitSFSparse() for details.}
}
\value{
An MFAIR object containing the information about the fitted MFAI model using greedy algorithm.
}
\description{
Fit the MFAI model using greedy algorithm.
}
