% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairSingleFactor.R
\name{fitSFFully}
\alias{fitSFFully}
\title{Fit the single factor MFAI model with fully observed main data matrix.}
\usage{
fitSFFully(
  Y,
  X,
  init,
  learning_rate,
  tree_parameters,
  iter_max = 5000,
  tol_stage1 = 0.1,
  tol_stage2 = 1e-05,
  verbose_sf = TRUE,
  verbose_loop = TRUE,
  save_tree_list = TRUE
)
}
\arguments{
\item{Y}{Main data matrix.}

\item{X}{A data.frame containing the auxiliary information.}

\item{init}{A MFAIRSingleFactor object containing the initial parameters for the single factor MAFI model.}

\item{learning_rate}{Numeric. Parameter for the gradient boosting part.}

\item{tree_parameters}{A list containing the parameters for the gradient boosting part.}

\item{iter_max}{Integer. Maximum iterations allowed.}

\item{tol_stage1}{Numeric. Convergence criterion in the first step.}

\item{tol_stage2}{Numeric. Convergence criterion in the first step.}

\item{verbose_sf}{Logical. Whether to display the detailed information.}

\item{verbose_loop}{Logical. Whether to display the detailed information when looping.}

\item{save_tree_list}{Logical. Whether to save the tree list.}
}
\value{
A MFAIRSingleFactor object containing the information about the fitted single factor MFAI model.
}
\description{
Fit the single factor MFAI model with fully observed main data matrix.
}
