% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairSingleFactor.R
\name{fitSFMissing}
\alias{fitSFMissing}
\title{Fit the single factor MFAI model.}
\usage{
fitSFMissing(
  Y,
  obs_indices,
  X,
  init,
  learning_rate,
  rpart_control,
  iter_max = 5000,
  tol_stage1 = 0.1,
  tol_stage2 = 1e-05,
  verbose_inner = TRUE,
  save_tree_list = TRUE
)
}
\arguments{
\item{Y}{Main data matrix.}

\item{X}{A data.frame containing the auxiliary information.}

\item{init}{A MFAIRSingleFactor object containing the initial parameters for the single factor MAFI model.}

\item{rpart_control}{A list containing the parameters for the gradient boosting part.}

\item{...}{see fitSFFully() and fitSFMissing()}
}
\value{
A MFAIRSingleFactor object containing the information about the fitted single factor MFAI model.
}
\description{
Fit the single factor MFAI model.
}
