% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairImportance.R
\name{getImportanceSF}
\alias{getImportanceSF}
\title{Get importance measures of auxiliary covariates in a single factor.}
\usage{
getImportanceSF(tree_list, variables_names)
}
\arguments{
\item{tree_list}{A fitted function represented by a list of trees.}

\item{variables_names}{The names of the auxiliary covariates.}
}
\value{
Importance score vector. Each entry is the importance score of one auxiliary covariate.
}
\description{
Get importance measures of auxiliary covariates in a single factor.
}
