% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairInitialize.R
\name{initFully}
\alias{initFully}
\title{Initialize the parameters for the single factor MAFI model with fully observed main data matrix.}
\usage{
initFully(Y)
}
\arguments{
\item{Y}{A fully observed main data matrix (no missing entries).}
}
\value{
MFAIRSingleFactor object containing the initial parameters for the single factor MAFI model.
}
\description{
Initialize the parameters for the single factor MAFI model with fully observed main data matrix.
}
