% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairInitialize.R
\name{initMissing}
\alias{initMissing}
\title{Initialize the parameters for the single factor MAFI model with partially observed main data matrix.}
\usage{
initMissing(Y)
}
\arguments{
\item{Y}{A partially observed main data matrix.}
}
\value{
MFAIRSingleFactor object containing the initial parameters for the single factor MAFI model.
}
\description{
Initialize the parameters for the single factor MAFI model with partially observed main data matrix.
}
