% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairInitialization.R
\name{initSF}
\alias{initSF}
\title{Initialize the parameters for the single factor MAFI model.}
\usage{
initSF(Y, Y_missing, Y_sparse, n_obs)
}
\arguments{
\item{Y}{Main data matrix.}

\item{Y_missing}{Logical. Whether the main data matrix is partially observed.}

\item{Y_sparse}{Logical. Whether the main data matrix is in sparse mode.}

\item{n_obs}{Integer. Total number of observed entries.}
}
\value{
MFAIRSingleFactor object containing the initial parameters for the single factor MAFI model.
}
\description{
Initialize the parameters for the single factor MAFI model.
}
