% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairObject.R
\docType{class}
\name{matrixORdgCMatrix-class}
\alias{matrixORdgCMatrix-class}
\title{Define the matrixORdgCMatrix class as the union of matrix and Matrix::dgCMatrix.}
\description{
Define the matrixORdgCMatrix class as the union of matrix and Matrix::dgCMatrix.
}
