% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairObject.R
\docType{class}
\name{MFAIR-class}
\alias{MFAIR-class}
\title{Each MFAIR object has a number of slots which store information. Key slots to access are listed below.}
\value{
MFAIR class.
}
\description{
Each MFAIR object has a number of slots which store information. Key slots to access are listed below.
}
\section{Slots}{

\describe{
\item{\code{Y}}{A matrix or Matrix::dgCMatrix. The main data matrix of N samples and M features.}

\item{\code{X}}{A data.frame. The auxiliary information data frame of N samples and C covariates.}

\item{\code{Y_sparse}}{Logical. Whether the main data matrix Y is stored in the sparse mode.}

\item{\code{Y_center}}{Logical. Whether the main data matrix Y is centered.}

\item{\code{Y_mean}}{Numeric. Mean of the main data matrix Y if centered. Zero if not.}

\item{\code{Y_missing}}{Logical. Whether the main data matrix Y is partially observed.}

\item{\code{n_obs}}{integer. Total number of observed entries in Y.}

\item{\code{N}}{An integer. Number of rows (samples) of Y, also the number of rows (samples) of X.}

\item{\code{M}}{An integer. Number of columns (features) of Y.}

\item{\code{C}}{An integer. Number of columns (auxiliary covariates) of X.}

\item{\code{K_max}}{An integer. Please note that increasing K_max does not ensure that the actual K also increases since K_max is just an upper bound, and the model will automatically infer K below K_max under the default setting. If you want to fit the model with larger rank K, please set the \code{null_check} argument as FALSE, or make sure that K_max is large enough and the \code{tol_snr} argument in the fitting function \code{fitGreedy()} is small enough simultaneously in the fitting function \code{fitGreedy()}.}

\item{\code{K}}{An integer. The inferred rank of Y.}

\item{\code{Z}}{A list of length K. Each element is a vector of length N, representing the posterior mean of one factor in the MFAI model.}

\item{\code{a_sq}}{A list of length K with each element representing the posterior variance of one factor in the MFAI model.  For fully observed Y, all N samples of one specific factor share the same posterior variance, then each element is a numeric value. For Y with missing data, the samples have different posterior variances, then each element is a vector of length N.}

\item{\code{W}}{A list of length K. Each element is a vector of length M, representing the posterior mean of one loading in the MFAI model.}

\item{\code{b_sq}}{A list of length K with each element representing the posterior variance of one loading in the MFAI model. For fully observed Y, all M features of one specific loading share the same posterior variance, then each element is a numeric value. For Y with missing data, the features have different posterior variances, then each element is a vector of length M.}

\item{\code{tau}}{A list of length K. Each elements is a numeric value representing the precision parameter for the combination of each pair of factor and loading.}

\item{\code{beta}}{A list of length K. Each elements is a numeric value representing the precision parameter for each factor Z_k.}

\item{\code{FX}}{A list of length K. Each element is a vector of length N, representing the prior mean of one factor in the MFAI model.}

\item{\code{tree_0}}{A list of length K. Each element is a numeric value representing tree_0 in one factor, which is defined as the mean of mu vector.}

\item{\code{tree_lists}}{A list of length K, containing K fitted functions and each function is represented as a list of trees, i.e., the k-th list corresponds to function F_k(.) in the MFAI model.}

\item{\code{initialization}}{A list. Initialization of the fitted model.}

\item{\code{learning_rate}}{Numeric. The learning rate in the gradient boosting part.}

\item{\code{tree_parameters}}{A list of options that control details of the rpart algorithm.}

\item{\code{project}}{Character. Name of the project (for record keeping).}
}}

