% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairObject.R
\docType{class}
\name{MFAIRSingleFactor-class}
\alias{MFAIRSingleFactor-class}
\title{MFAIRSingleFactor object contains the key information about the fitted single factor MFAI model.}
\value{
MFAIRSingleFactor class.
}
\description{
MFAIRSingleFactor object contains the key information about the fitted single factor MFAI model.
}
\section{Slots}{

\describe{
\item{\code{Y_missing}}{Logical. Whether the main data matrix Y is partially observed.}

\item{\code{n_obs}}{Integer. Total number of observed entries in Y.}

\item{\code{mu}}{An vector of length N representing the inferred loading, corresponding to the posterior mean of z in the single factor MFAI model.}

\item{\code{a_sq}}{Numeric. The posterior variance of the loading z. For fully observed Y, all N elements of the loading share the same posterior variance, then a_sq is a single number. For Y with missing data, the elements have different posterior variances, then a_sq is a vector of length N.}

\item{\code{nu}}{An vector of length M representing the inferred factor, corresponding to the posterior mean of w in the single factor MFAI model.}

\item{\code{b_sq}}{Numeric. The posterior variance of the factor w. For fully observed Y, all M elements of the factor share the same posterior variance, then b_sq is a single number. For Y with missing data, the elements have different posterior variances, then b_sq is a vector of length M.}

\item{\code{tau}}{Numeric. Precision parameter this pair of loading/factor.}

\item{\code{beta}}{Numeric. Precision parameter for this loading z.}

\item{\code{FX}}{An vector of length N representing the prior mean of z, corresponding to F(X) in the single factor MFAI model.}

\item{\code{tree_list}}{A list containing multiple decision trees, corresponding to function F(.) in the single factor MFAI model.}

\item{\code{project}}{Character. Name of the project (for record keeping).}
}}

