% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairPredict.R
\name{predict,MFAIR-method}
\alias{predict,MFAIR-method}
\title{Prediction function for MFAIR object.}
\usage{
\S4method{predict}{MFAIR}(object)
}
\arguments{
\item{MFAIR}{A model object for which prediction is desired.}
}
\value{
Predicted matrix with the same dimension as that of Y.
}
\description{
Prediction function for MFAIR object.
}
\examples{
mfairObject <- createMFAIR(Y, X)
mfairObject <- fitGreedy(mfairObject)
Y_hat <- predict(mfairObject)
}
