% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairPredict.R
\name{predict,MFAIRSingleFactor-method}
\alias{predict,MFAIRSingleFactor-method}
\title{Predict function for MFAIRSingleFactor object.}
\usage{
\S4method{predict}{MFAIRSingleFactor}(object)
}
\arguments{
\item{MFAIRSingleFactor}{MFAIRSingleFactor object.}
}
\value{
Predicted matrix with the same dimension as that of Y.
}
\description{
Predict function for MFAIRSingleFactor object.
}
