% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairPredict.R
\name{predictFX}
\alias{predictFX}
\title{Prediction function for fitted functions.}
\usage{
predictFX(object, newdata, which_factors = seq_len(object@K))
}
\arguments{
\item{object}{MFAIR object.}

\item{newdata}{Data frame containing the values at which predictions are required.}

\item{which_factors}{Which factors, i.e., which fitted functions are used. All K factors are used by default.}
}
\value{
A matrix containing predicted F(X). Each row is a new sample and each column is a factor.
}
\description{
Prediction function for fitted functions.
}
