% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairPredict.R
\name{predictFXSF}
\alias{predictFXSF}
\title{Prediction function for fitted function F() in single factor.}
\usage{
predictFXSF(tree_list, tree_0, newdata, learning_rate)
}
\arguments{
\item{tree_list}{A fitted function represented by a list of trees.}

\item{tree_0}{A numeric value representing the mean value.}

\item{newdata}{Data frame containing the values at which predictions are required.}

\item{learning_rate}{Numeric. The learning rate in the gradient boosting part.}
}
\value{
A vector containing predicted F(X). Each entry corresponds to a new sample.
}
\description{
Prediction function for fitted function F() in single factor.
}
