% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairUtils.R
\name{projSparse}
\alias{projSparse}
\title{Project a matrix with given indices and store the result in the sparse mode.}
\usage{
projSparse(Y, obs_indices)
}
\arguments{
\item{Y}{A matrix to be projected.}

\item{obs_indices}{A matrix containing 1-based indices of the observed entries in the matrix Y. The first column represents row and the second column represents column.}
}
\value{
A dgCMatrix containing the projection result.
}
\description{
Project a matrix with given indices and store the result in the sparse mode.
}
