% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softImputeCV.R
\name{softImputeCrossVal}
\alias{softImputeCrossVal}
\title{Cross-validation for softImpute}
\usage{
softImputeCrossVal(
  Y,
  rank_max = NULL,
  lambda_range = NULL,
  nfold = 10,
  para_length = 100
)
}
\arguments{
\item{Y}{A matrix. The main data matrix of N samples and M features.}

\item{rank_max}{An integer. The maximum rank allowed.}

\item{lambda_range}{A vector containing the minimal and maximal value of the parameter lambda.}

\item{nfold}{An integer. The total number of validation sets.}

\item{para_length}{An integer. The total number of parameter lambda.}
}
\value{
A list containing the cross-validation results.
}
\description{
Cross-validation for softImpute
}
