% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairUpdate.R
\name{updateMFAIR}
\alias{updateMFAIR}
\title{Update the k-th factor of the MFAIR object in the backfitting algorithm.}
\usage{
updateMFAIR(object, object_sf, k)
}
\arguments{
\item{object}{An MFAIR object containing the initial information about the K-factor MFAI model.}

\item{object_sf}{A MFAIRSingleFactor containing the information about the newly fitted single factor MFAI model.}

\item{k}{Integer. Which fator to be updated.}
}
\value{
MFAIR object containing the information about the new fitted single factor.
}
\description{
Update the k-th factor of the MFAIR object in the backfitting algorithm.
}
