% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairObject.R
\name{createMFAIR}
\alias{createMFAIR}
\title{Create the MFAIR object with main data matrix and auxiliary information.}
\usage{
createMFAIR(Y, X, K_max = 1L, project = "MFAIR")
}
\arguments{
\item{Y}{A matrix. The main data matrix of N samples and M features.}

\item{X}{A data.frame. The auxiliary information data frame of N samples and C covariates.}

\item{K_max}{An integer. The maximum rank allowed in the model.}

\item{project}{Character. Name of the project (for record keeping).}
}
\value{
Returns MFAIR object, with main data matrix and auxiliary information.
}
\description{
Create the MFAIR object with main data matrix and auxiliary information.
}
