% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairFitting.R
\name{fitGreedy}
\alias{fitGreedy}
\title{Create generic for greedy algorithm.}
\usage{
fitGreedy(object, ...)
}
\arguments{
\item{object}{a model object for which greedy algorithm is applied.}

\item{...}{Additional arguments.}
}
\value{
MFAIR object containing the information about the fitted MFAI model using greedy algorithm.
}
\description{
Create generic for greedy algorithm.
}
