% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairELBO.R
\name{getELBO}
\alias{getELBO}
\title{Compute the evidence lower bound (ELBO) for fitted single factor MFAI model.}
\usage{
getELBO(Y, object)
}
\arguments{
\item{Y}{Observed main data matrix.}

\item{object}{MFAIRSingleFactor object containing the information about the fitted single factor MFAI model.}
}
\value{
Numeric. The ELBO.
}
\description{
Compute the evidence lower bound (ELBO) for fitted single factor MFAI model.
}
