% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairImportance.R
\name{getImportance}
\alias{getImportance}
\title{Get importance measures of auxiliary covariates.}
\usage{
getImportance(object, which_factors = c(1:object@K))
}
\arguments{
\item{object}{MFAIR object.}

\item{which_factors}{Which factors, i.e., which fitted functions will be evaluated. All K factors are evaluated by default.}
}
\value{
Importance score matrix. Each row is an auxiliary covariate and each column is a factor.
}
\description{
Get importance measures of auxiliary covariates.
}
\details{
The \code{rpart::rpart()} function will automatically change special characters in variable names to a dot which may cause some inconsistency errors.
}
