% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairInitialize.R
\name{initSF}
\alias{initSF}
\title{Initialize the parameters for the single factor MAFI model.}
\usage{
initSF(Y, Y_missing, n_obs)
}
\arguments{
\item{Y}{Main data matrix.}
}
\value{
MFAIRSingleFactor object containing the initial parameters for the single factor MAFI model.
}
\description{
Initialize the parameters for the single factor MAFI model.
}
\section{Slots}{

\describe{
\item{\code{Y_missing}}{Logical. Whether the main data matrix Y is partially observed.}

\item{\code{n_obs}}{Integer. Total number of observed entries in Y.}
}}

