% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairObject.R
\docType{class}
\name{MFAIR-class}
\alias{MFAIR-class}
\title{Each MFAIR object has a number of slots which store information. Key slots to access are listed below.}
\value{
MFAIR class.
}
\description{
Each MFAIR object has a number of slots which store information. Key slots to access are listed below.
}
\section{Slots}{

\describe{
\item{\code{Y}}{numeric. The main data matrix of N samples and M features.}

\item{\code{X}}{data.frame. The auxiliary information data frame of N samples and C covariates.}

\item{\code{Y_missing}}{logical. Whether the main data matrix Y is partially observed.}

\item{\code{K_max}}{integer. The maximum rank allowed in the model.}

\item{\code{K}}{integer. The inferred rank of Y.}

\item{\code{Z}}{numeric. Estimated loading matrix with dimension N * K, corresponding to the inferred posterior mean of Z in the MFAI model.}

\item{\code{W}}{numeric. Estimated factor matrix with dimension M * K, corresponding to the inferred posterior mean of W in the MFAI model.}

\item{\code{tau}}{numeric. A vector of length K, containing the precision parameter for each pair of loading/factor.}

\item{\code{beta}}{numeric. A vector of length K, containing the precision parameter for each loading Z_k.}

\item{\code{FX}}{numeric. An N * K matrix representing the prior mean of Z, corresponding to F(X) in the MFAI model.}

\item{\code{tree_lists}}{list. A list containing K fitted tree lists, each corresponding to function F_k(.) in the MFAI model.}

\item{\code{initialization}}{list. Initialization of the fitted model.}

\item{\code{boosting_parameters}}{list. A list of options that control details of the rpart algorithm.}

\item{\code{project}}{character. Name of the project (for record keeping).}

\item{\code{.}}{}
}}

