% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairPredict.R
\name{predict,MFAIR-method}
\alias{predict,MFAIR-method}
\title{Prediction function for MFAIR object.}
\usage{
\S4method{predict}{MFAIR}(object, which_factors = seq_len(object@K))
}
\arguments{
\item{object}{A model object for which prediction is desired.}

\item{which_factors}{Which factors, i.e., which columns of Z and W, are used to make prediction. All K factors are used by default.}
}
\value{
Predicted matrix with the same dimension as that of Y.
}
\description{
Prediction function for MFAIR object.
}
