% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfairPredict.R
\name{predict,MFAIRSingleFactor-method}
\alias{predict,MFAIRSingleFactor-method}
\title{Prediction function for MFAIRSingleFactor object.}
\usage{
\S4method{predict}{MFAIRSingleFactor}(object)
}
\arguments{
\item{object}{A model object for which prediction is desired.}
}
\value{
Predicted matrix with the same dimension as that of Y.
}
\description{
Prediction function for MFAIRSingleFactor object.
}
